Add logic compiler tool
The logic compiler tool is a separate tool that reads all files written in the
".w5logic" format and generates C# code the represents that.

The generated code is a module similar to the previous theme extensions and
provide game modes, character roles and more that can be used to extend the
server.