UI: Migrate core components to use the new language file system
Some paths have been updated and those have to be reflected in the UI to still
show the correct strings. Also is the index format extended and this has to be
reflected in the UI as well.

The language object is now hidden from the `Debug.toString` representation to
make it smaller and easier to focus on important parts.